requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==38.0.1
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.21
pycryptodome==3.20.0
Werkzeug==3.0.1
